Declare radius_xlat return as size_t

+++ src/main/radwho.c
@@ -80,9 +80,9 @@ int log_stripped_names;
 struct main_config_t mainconfig;
 char *request_log_file = NULL;
 char *debug_log_file = NULL;
-int radius_xlat(char *out, UNUSED int outlen, UNUSED const char *fmt,
-		UNUSED REQUEST *request,
-		UNUSED RADIUS_ESCAPE_STRING func, UNUSED void *arg)
+size_t radius_xlat(char *out, UNUSED int outlen, UNUSED const char *fmt,
+		   UNUSED REQUEST *request,
+		   UNUSED RADIUS_ESCAPE_STRING func, UNUSED void *arg)
 {
 	*out = 0;
 	return 0;
